Make read_particlefile public
