+++ tests/utils.py
@@ -8,16 +8,12 @@ from pathlib import Path
 
 import cftime
 import numpy as np
-import pandas as pd
-import pyarrow as pa
-import pyarrow.parquet as pq
 import xarray as xr
 
 import parcels
 from parcels import FieldSet, Particle, Variable
 from parcels._core.xgrid import _FIELD_DATA_ORDERING, XGrid, get_axis_from_dim_name
 from parcels._datasets.structured.generated import simple_UV_dataset
-from parcels._typing import PathLike
 
 PROJECT_ROOT = Path(__file__).resolve().parents[1]
 TEST_ROOT = PROJECT_ROOT / "tests"
@@ -161,48 +157,10 @@ def round_and_hash_float_array(arr, decimals=6):
 def assert_cftime_like_particlefile(parquet_path: Path) -> None:
     assert parquet_path.suffix == ".parquet", "Path must be a parquet file"
 
-    df = read_particlefile(parquet_path, decode_times=True)
+    df = parcels.read_particlefile(parquet_path, decode_times=True)
 
     # check first value (and hence rest of array) is what we expect
     assert isinstance(df["time"].values[0], (cftime.datetime, np.datetime64)), (
         "CF-time values in Parquet did not get properly decoded. Are the attributes correct?"
     )
     return
-
-
-def read_particlefile(path: PathLike, decode_times: bool = True) -> pd.DataFrame:
-    path = Path(path)
-
-    assert path.suffix == ".parquet", "Only Parquet files are supported"
-
-    table = pq.read_table(path)
-
-    try:
-        time_field = table.field("time")
-    except KeyError as e:
-        raise ValueError(
-            f"Could not find 'time' column in parquet file. Are you sure {path=!r} is a particlefile?"
-        ) from e
-
-    assert pa.types.is_floating(time_field.type) or pa.types.is_integer(time_field.type), (
-        f"'time' column must be numeric, got {time_field.type}"
-    )
-
-    try:
-        assert b"units" in time_field.metadata
-    except AssertionError as e:
-        raise ValueError(f"Could not find 'units' in the 'time' column metadata for parquet {path=!r}.") from e
-
-    attrs = {k.decode(): v.decode() for k, v in time_field.metadata.items()}
-
-    df = pd.read_parquet(path)
-    if not decode_times:
-        return df
-
-    values = table.column("time").to_numpy()
-    var = xr.Variable(("time",), values, attrs)
-    values = xr.coders.CFDatetimeCoder(time_unit="s").decode(var).values
-
-    df["time"] = values
-
-    return df
